Added more functions in Tile.lua
Tiles in TileTable from the Play Button Scene.lua now have clicked event
and a hover event implemented. Click event is active within the code but
hover event is locked behind comments. When you click on a tile from the
map now in the play button scene, the console should display whether or
not the tile has been pressed.

Storing players within a tile will be later implemented because of the
complexity required when entering and leaving a specific tile. Also
highlighting tile will also be later implemented because of the
limitations of needing the same image of a panel but with a transparent
color to easily and visually indicate that it is being hovered over by a
mouse.