Theoretically finished TileMaps
Added tile functionality to add and remove players from a tile using a TileMaps.lua that in a way acts as a C++ maps reference. TileMaps.lua has an array of keys that stores all of a specific tile location in a map, and an array of data type that will contain another array to store players that land on that tile location. Both key array and datatype array are parallel to each other so information stays consistent.

This tile function does have a way to be tested yet until further PlayerGUI code is created .